Workaround for cargo not running tests in examples folder.
